Move DEBUG code paths from the main sceneDelegate to a new DebuggingSceneDelegate subclass to separate debug and test-specific code from production code
